Allow for custom analyzer factory
